+++ src/index.html
@@ -31,6 +31,7 @@
 
     <script type="module" src="/main.js"></script>
     <script type="module" src="/js/header-menu.js"></script>
+    <script type="module" src="/js/feedback.js"></script>
     <script type="module" src="/js/accordion.js"></script>
   </body>
 </html>
